addtripdestination component added with functionality to add/delete locations
