+++ src/components/destination/AddTripDestination.js
@@ -1,17 +1,20 @@
 import { useState, useEffect, useRef } from "react"
-import { getDestinations, addDestination, deleteDestination } from "../../managers/DestinationManager"
-import { updateTrip, getSingleTrip, getDestinationByTrip, addTripDestination } from "../../managers/TripManager"
+import { addDestination, deleteDestination } from "../../managers/DestinationManager"
+import { getSingleTrip, addTripDestination } from "../../managers/TripManager"
+import { getStatus } from "../../managers/StatusManager"
 import { FaTrashAlt } from "react-icons/fa";
 
-export const AddTripDestination = ({tripId, currentTrip, setCurrentTrip}) => {
+export const AddTripDestination = ({ tripId, currentTrip, setCurrentTrip, status }) => {
 
     const locationRef = useRef()
     const stateRef = useRef()
     const latRef = useRef()
     const longRef = useRef()
     const tipsRef = useRef()
+    const destStatusRef = useRef()
     const [refresh, setRefresh] = useState(false)
     const [destinations, setDestinations] = useState([])
+
     const [newDestination, setNewDestination] = useState({
         id: 0,
         location: "",
@@ -19,7 +22,7 @@ export const AddTripDestination = ({tripId, currentTrip, setCurrentTrip}) => {
         latitude: 0,
         longitude: 0,
         tips: "",
-        status: 0
+        destination_status: 0
     })
 
     const handleNewDestinationInfo = (event) => {
@@ -34,6 +37,12 @@ export const AddTripDestination = ({tripId, currentTrip, setCurrentTrip}) => {
         setNewDestination(latLong)
     }
 
+    const parseDestStatus = (event) => {
+        const destination_status = Object.assign({}, newDestination)
+        destination_status[event.target.name] = parseInt(event.target.value)
+        setNewDestination(destination_status)
+    }
+
     const createNewDestination = (event) => {
         event.preventDefault()
 
@@ -42,7 +51,8 @@ export const AddTripDestination = ({tripId, currentTrip, setCurrentTrip}) => {
             state: stateRef.current.value,
             latitude: parseFloat(latRef.current.value),
             longitude: parseFloat(longRef.current.value),
-            tips: tipsRef.current.value
+            tips: tipsRef.current.value,
+            destination_status: parseInt(destStatusRef.current.value)
         }
         setDestinations([...destinations, newDestination])
         locationRef.current.value = ''
@@ -50,6 +60,7 @@ export const AddTripDestination = ({tripId, currentTrip, setCurrentTrip}) => {
         latRef.current.value = ''
         longRef.current.value = ''
         tipsRef.current.value = ''
+        destStatusRef.current.value = ''
 
         addDestination(newDestination)
             .then((destination) => {
@@ -146,6 +157,33 @@ export const AddTripDestination = ({tripId, currentTrip, setCurrentTrip}) => {
                             onChange={handleNewDestinationInfo}
                         />
                     </div>
+                    <div className="newStatusSelect">
+                        <select
+                            name="destination_status"
+                            className="input"
+                            ref={destStatusRef}
+                            value={newDestination.destination_status}
+                            onChange={parseDestStatus}
+                        // onChange={(event) => {
+                        //     const copy = { ...newDestination }
+                        //     copy.statusId = parseInt(event.target.value)
+                        //     setNewItem(copy)
+                        // }}
+                        >
+                            <option value="0">Status Select</option>
+                            {
+                                status.map(s => (
+                                    <option
+                                        key={`status--${s.id}`}
+                                        value={s.id}>
+                                        {s.type}
+                                    </option>
+                                ))
+                            }
+                        </select>
+                    </div>
+
+
 
 
                     <button className="button is-small addDestinationButton"
